Add default implementations for plugs
