--- Source/Mosa.Demo.SVGAWorld.x86/Plugs/ConsolePlug.cs
@@ -1,82 +1,0 @@
-// Copyright (c) MOSA Project. Licensed under the New BSD License.
-
-using System;
-using Mosa.Kernel.x86;
-using Mosa.Runtime.Plug;
-
-namespace Mosa.Demo.SVGAWorld.x86.Plugs;
-
-public static class ConsolePlug
-{
-	[Plug("System.Console::ResetColor")]
-	internal static void ResetColor()
-	{
-		Screen.Color = (byte)ConsoleColor.White;
-		Screen.BackgroundColor = (byte)ConsoleColor.Black;
-	}
-
-	[Plug("System.Console::Clear")]
-	internal static void Clear()
-	{
-		Screen.Clear();
-	}
-
-	[Plug("System.Console::WriteLine")]
-	internal static void WriteLine()
-	{
-		Serial.Write(Serial.COM1, "\r\n");
-		Screen.WriteLine();
-	}
-
-	[Plug("System.Console::WriteLine")]
-	internal static void WriteLine(string value)
-	{
-		Serial.Write(Serial.COM1, value);
-		Serial.Write(Serial.COM1, "\r\n");
-		Screen.WriteLine(value);
-	}
-
-	[Plug("System.Console::Write")]
-	internal static void Write(string value)
-	{
-		Serial.Write(Serial.COM1, value);
-		Screen.Write(value);
-	}
-
-	[Plug("System.Console::Write")]
-	internal static void Write(char value)
-	{
-		Serial.Write(Serial.COM1, (byte)value);
-		Screen.Write(value);
-	}
-
-	[Plug("System.Console::SetCursorPosition")]
-	internal static void SetCursorPosition(int left, int top)
-	{
-		Screen.Goto((uint)left, (uint)top);
-	}
-
-	[Plug("System.Console::GetForegroundColor")]
-	internal static ConsoleColor GetForegroundColor()
-	{
-		return (ConsoleColor)Screen.Color;
-	}
-
-	[Plug("System.Console::GetBackgroundColor")]
-	internal static ConsoleColor GetBackgroundColor()
-	{
-		return (ConsoleColor)Screen.BackgroundColor;
-	}
-
-	[Plug("System.Console::SetForegroundColor")]
-	internal static void SetForegroundColor(ConsoleColor color)
-	{
-		Screen.Color = (byte)color;
-	}
-
-	[Plug("System.Console::SetBackgroundColor")]
-	internal static void SetBackgroundColor(ConsoleColor color)
-	{
-		Screen.BackgroundColor = (byte)color;
-	}
-}
